feat: Set up users table/schema
 - Create the users table and populate it with some dummy data
 - Set up the users schema using knex

 Next I need to update the ContactsAPI to only do CRUD operations on the correct contacts, given the current user.
